requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
PyGithub==2.1.1
PyJWT==2.8.0
cryptography==41.0.7
python-dotenv==1.0.0
httpx==0.26.0

